+++ ormar/relations/alias_manager.py
@@ -34,12 +34,27 @@ class AliasManager:
 
     def __init__(self) -> None:
         self._aliases_new: Dict[str, str] = dict()
+        self._reversed_aliases: Dict[str, str] = dict()
 
     def __contains__(self, item: str) -> bool:
         return self._aliases_new.__contains__(item)
 
     def __getitem__(self, key: str) -> Any:
         return self._aliases_new.__getitem__(key)
+
+    @property
+    def reversed_aliases(self) -> Dict:
+        """
+        Returns swapped key-value pairs from aliases where alias is the key.
+
+        :return: dictionary of prefix to relation
+        :rtype: Dict
+        """
+        if self._reversed_aliases:
+            return self._reversed_aliases
+        reversed_aliases = {v: k for k, v in self._aliases_new.items()}
+        self._reversed_aliases = reversed_aliases
+        return self._reversed_aliases
 
     @staticmethod
     def prefixed_columns(